+++ Grid/Source/ColumnSourceInfo.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Dtc\GridBundle\Grid\Source;
+
+class ColumnSourceInfo
+{
+    public $columns;
+    public $idColumn;
+    public $sort;
+}