+++ Manager/GridSourceManager.php
@@ -8,6 +8,8 @@ use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityManagerInterface;
+use Dtc\GridBundle\Grid\Source\ColumnSource;
+use Dtc\GridBundle\Grid\Source\ColumnSourceInfo;
 use Dtc\GridBundle\Grid\Source\DocumentGridSource;
 use Dtc\GridBundle\Grid\Source\EntityGridSource;
 use Dtc\GridBundle\Grid\Source\GridSourceInterface;
@@ -18,8 +20,6 @@ class GridSourceManager
     protected $sourcesByName;
 
     protected $reader;
-    protected $cacheDir;
-    protected $debug;
 
     /** @var AbstractManagerRegistry */
     protected $registry;
@@ -31,6 +31,7 @@ class GridSourceManager
 
     protected $extraGridSources;
 
+    protected $columnSource;
     /**
      * @var array|null Null means all entities allowed, empty array means no entities allowed
      */
@@ -38,23 +39,24 @@ class GridSourceManager
 
     /**
      * GridSourceManager constructor.
-     *
-     * @param string $cacheDir
-     * @param bool   $debug
      */
-    public function __construct(Reader $reader, $cacheDir, $debug = false)
+    public function __construct(ColumnSource $columnSource)
     {
-        $this->cacheDir = $cacheDir;
-        $this->reader = $reader;
+        $this->columnSource = $columnSource;
         $this->reflectionAllowedEntities = [];
-        $this->debug = $debug;
         $this->sources = array();
+    }
+
+    public function setReader(Reader $reader)
+    {
+        $this->reader = $reader;
     }
 
     /**
      * @param array|string $allowedEntities Array of allowed entities or string '*' to allow all entities. Use empty array to specify no entities allowed for reflection.
      */
-    public function setReflectionAllowedEntities($allowedEntities) {
+    public function setReflectionAllowedEntities($allowedEntities)
+    {
         $this->reflectionAllowedEntities = is_array($allowedEntities) ? array_flip($allowedEntities) : ('*' === $allowedEntities ? null : []);
     }
 
@@ -68,22 +70,6 @@ class GridSourceManager
         $this->mongodbRegistry = $registry;
     }
 
-    /**
-     * @return string
-     */
-    protected function getCacheDir()
-    {
-        return $this->cacheDir;
-    }
-
-    /**
-     * @return bool
-     */
-    protected function getDebug()
-    {
-        return $this->debug;
-    }
-
     public function add($id, GridSourceInterface $gridSource)
     {
         $this->extraGridSources[$id] = $gridSource;
@@ -92,8 +78,11 @@ class GridSourceManager
     /**
      * @param ObjectManager $manager
      * @param $entityOrDocument
+     *z
      *
      * @return DocumentGridSource|EntityGridSource|null
+     *
+     * @throws \Exception
      */
     protected function getGridSource($manager, $entityOrDocument)
     {
@@ -101,30 +90,48 @@ class GridSourceManager
         if ($repository) {
             $className = $repository->getClassName();
             $classMetadata = $manager->getClassMetadata($className);
+            $params = [$manager, $entityOrDocument, null === $this->reflectionAllowedEntities || isset($this->reflectionAllowedEntities[$entityOrDocument])];
+            if ($this->reader) {
+                $params[] = $this->reader;
+            }
+            $columnSourceInfo = call_user_func_array([$this->columnSource, 'getColumnSourceInfo'], $params);
             $name = $classMetadata->getName();
-            $reflectionClass = $classMetadata->getReflectionClass();
-            $annotation = $this->reader->getClassAnnotation($reflectionClass, 'Dtc\GridBundle\Annotation\Grid');
-            if (!$annotation && !isset($this->reflectionAllowedEntities[$entityOrDocument]) && null !== $this->reflectionAllowedEntities) {
-                throw new \Exception("GridSource requested for '$entityOrDocument' but no Grid annotation found");
+            if ($columnSourceInfo) {
+                return $this->getGridSourceFromColumnSourceInfo($manager, $className, $name, $columnSourceInfo);
             }
-            if ($manager instanceof EntityManagerInterface) {
-                $gridSource = new EntityGridSource($manager, $name);
-            } else {
-                $gridSource = new DocumentGridSource($manager, $name);
-            }
-            $gridSource->setAnnotationReader($this->reader);
-            $gridSource->setCacheDir($this->cacheDir);
-
-            $gridSource->setDebug($this->debug);
-            $gridSource->autoDiscoverColumns();
-            $this->sourcesByName[$name] = $gridSource;
-            $this->sourcesByClass[$className] = $gridSource;
-            $gridSource->setId($className);
-
-            return $gridSource;
         }
 
         return null;
+    }
+
+    /**
+     * @param $manager
+     * @param $className
+     * @param $name
+     * @param ColumnSourceInfo $columnSourceInfo
+     *
+     * @return DocumentGridSource|EntityGridSource
+     *
+     * @throws \Exception
+     */
+    private function getGridSourceFromColumnSourceInfo($manager, $className, $name, ColumnSourceInfo $columnSourceInfo)
+    {
+        if ($manager instanceof EntityManagerInterface) {
+            $gridSource = new EntityGridSource($manager, $name);
+        } else {
+            if (!$manager instanceof DocumentManager) {
+                throw new \Exception('Unknown ObjectManager type: '.get_class($manager));
+            }
+            $gridSource = new DocumentGridSource($manager, $name);
+        }
+        $gridSource->setIdColumn($columnSourceInfo->idColumn);
+        $gridSource->setColumns($columnSourceInfo->columns);
+        $this->sourcesByName[$name] = $gridSource;
+        $this->sourcesByClass[$className] = $gridSource;
+        $gridSource->setId($className);
+        $gridSource->setDefaultSort($columnSourceInfo->sort);
+
+        return $gridSource;
     }
 
     /**