+++ Tests/Manager/GridSourceManagerTest.php
@@ -3,6 +3,7 @@
 namespace Dtc\GridBundle\Tests\Manager;
 
 use Doctrine\Common\Annotations\AnnotationReader;
+use Dtc\GridBundle\Grid\Source\ColumnSource;
 use Dtc\GridBundle\Manager\GridSourceManager;
 use Dtc\GridBundle\Tests\Grid\Source\TestGridSource;
 use PHPUnit\Framework\TestCase;
@@ -14,7 +15,7 @@ class GridSourceManagerTest extends TestCase
     {
         $container = new Container();
         $container->setParameter('dtc_grid.custom_managers', []);
-        $gridSourceManager = new GridSourceManager(new AnnotationReader(), '/tmp', true);
+        $gridSourceManager = new GridSourceManager(new ColumnSource('/tmp', true));
         self::assertNotNull($gridSourceManager->all());
         self::assertInternalType('array', $gridSourceManager->all());
         self::assertEmpty($gridSourceManager->all());
@@ -24,7 +25,7 @@ class GridSourceManagerTest extends TestCase
     {
         $container = new Container();
         $container->setParameter('dtc_grid.custom_managers', []);
-        $gridSourceManager = new GridSourceManager(new AnnotationReader(), '/tmp', true);
+        $gridSourceManager = new GridSourceManager(new ColumnSource('/tmp', true));
         $gridSource = new TestGridSource();
         $gridSourceManager->add('test_grid_source', $gridSource);
         self::assertSame($gridSource, $gridSourceManager->get('test_grid_source'));
@@ -39,7 +40,7 @@ class GridSourceManagerTest extends TestCase
     {
         $container = new Container();
         $container->setParameter('dtc_grid.custom_managers', []);
-        $gridSourceManager = new GridSourceManager(new AnnotationReader(), '/tmp', true);
+        $gridSourceManager = new GridSourceManager(new ColumnSource('/tmp', true));
         $gridSource = new TestGridSource();
         try {
             $gridSourceManager->get('test_grirce');