+++ Util/CamelCase.php
@@ -2,20 +2,18 @@
 
 namespace Dtc\GridBundle\Util;
 
-trait CamelCaseTrait
+class CamelCase
 {
     /**
      * @param string $str
      *
      * @return string
      */
-    protected function fromCamelCase($str)
+    public static function fromCamelCase($str)
     {
-        $func = function ($str) {
+        $value = preg_replace_callback('/([A-Z])/', function ($str) {
             return ' '.$str[0];
-        };
-
-        $value = preg_replace_callback('/([A-Z])/', $func, $str);
+        }, $str);
         $value = ucfirst($value);
 
         return trim($value);